requirements.txt
celery==5.0.3
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
psycopg2==2.8.6